// src/modules/database-driver/connector/grpc/gen/GrpcEnums_pb.ts
export enum GrpcPriceInnerRecordHandling {
  /**
   * No special strategy set. Inner record id is not taken into account at all.
   *
   * @generated from enum value: NONE = 0;
   */
  NONE = 0,

  /**
   * Prices with same inner entity id will be sorted descending by priority value and first one
   * (i.e. the one with the biggest priority) will be used (others won't be considered at all)
   *
   * @generated from enum value: LOWEST_PRICE = 1;
   */
  LOWEST_PRICE = 1,

  /**
   * Prices with same inner entity id will be added up to a new computed aggregated price, prices must share same
   * tax rate percentage, currency and price list id in order to be added up
   *
   * @generated from enum value: SUM = 2;
   */
  SUM = 2,

  /**
   * Price handling mode that is used in cases when the information has not been fetched along with entity, and
   * is therefore unknown (even if some strategy is associated with the entity in reality).
   *
   * @generated from enum value: UNKNOWN = 3;
   */
  UNKNOWN = 3,
}
// ...
export enum GrpcCardinality {
  /**
   * No cardinality specified.
   *
   * @generated from enum value: NOT_SPECIFIED = 0;
   */
  NOT_SPECIFIED = 0,

  /**
   * Relation may be missing completely, but if it exists - there is never more than single relation of this type.
   *
   * @generated from enum value: ZERO_OR_ONE = 1;
   */
  ZERO_OR_ONE = 1,

  /**
   * There is always single relation of this type.
   *
   * @generated from enum value: EXACTLY_ONE = 2;
   */
  EXACTLY_ONE = 2,

  /**
   * Relation may be missing completely, but there may be also one or more relations of this type.
   *
   * @generated from enum value: ZERO_OR_MORE = 3;
   */
  ZERO_OR_MORE = 3,

  /**
   * There is always at least one relation of this type, but there may be also more than one.
   *
   * @generated from enum value: ONE_OR_MORE = 4;
   */
  ONE_OR_MORE = 4,
}
// ...
export enum GrpcEntityScope {
  /**
   * Entities that are currently active and reside in the live data set block.
   *
   * @generated from enum value: SCOPE_LIVE = 0;
   */
  SCOPE_LIVE = 0,

  /**
   * Entities that are no longer active and reside in the archive block.
   *
   * @generated from enum value: SCOPE_ARCHIVED = 1;
   */
  SCOPE_ARCHIVED = 1,
}